Factor out parse-extracted-messages.ts.
